First chunk of code for new ExplicitModelAdapter.
The new adapter will generate an intermediate datastructure that holds all transitions to be inserted.
This will combine the two phases (computing the state space and actually generating the matrix) and can also be used for dtmc and mdp models.

This datastructure is only a list of maps for each state. Each map represents a nondeterministic choice and maps target node ids to their probability.