Cleaned up handling of --verbose, proposing correct use of log levels from now on...
FATAL_LOG_LEVEL: Use, if we are going to crash.
ERROR_LOG_LEVEL: Use, if there is no reasonable way to continue.
WARN_LOG_LEVEL: Use, if we got something the average user should read.
INFO_LOG_LEVEL: Use, if this might in some cases be of interest.
DEBUG_LOG_LEVEL: Use, if this should usually not be relevant to a user.
TRACE_LOG_LEVEL: Use only during development.

There are three levels of verbosity:
- default: WARN and above
- verbose: INFO and above
- debug: DEBUG and above