Refactor dynamic reordering activation and provide inhibitor.
Provides an inhibitor mechanism:
Request a DynamicReorderingInhibitor from the DdManager. As long as that object
(wrapped in a unique_ptr) is in scope, no further reordering can occur.

For Sylvan, tweak reordering handling: Only throw an exception if
reordering is actually activated. Provide a no-op inhibitor, as
reordering is currently not supported by Sylvan.