Workaround for IntelTBB linker issue by CMake/Regex magic.
IntelTBB does not use symlinks (as commonly used) to reference its libraries but instead uses linker scripts.
These linker scripts do not work with GCC and linking fails.
As a workaround we manually set the correct library in CMake after extracting the path from the linker script with regex magic.

This workaround is highly hackish and might break in the future.