(LTL) Refactor unary/boolean state formulas to allow path formulas as well
Previously, the formula parser only supported AND, OR and NOT over state formulas.
For LTL, we need those over path formulas.

We tweak the grammar, the AST types and visitors, to also handle
UnaryBooleanPathFormula and BinaryBooleanPathformula. During parsing,
we determine on-demand whether to generate the path or state formula
variant by looking at the subformulas - if they are state formulas,
the boolean operator will be a state formula as well.