Expression Parser: Fixed backtracking mode.
With backtracking enabled, the expression parser is not supposed to throw an exception  on a formula like `1 & true`.
However, when such an error is encountered the expression parser currently overwrites the result parsed so far with a dummy expression which results in a BAD_ACCESS signal. This commit fixes the issue.