Added output of MEC statistics.
When computing reachability probabilities or expected total reachability rewards, the MEC quotient of the "maybestates"-subsystem might be computed (either because it is required by the used algorithm or because the user has set the `--force-require-unique`.
If this is the case and the `--verbose` has been set, some interesting statistics on the MECs will now be printed.