Fixes for transient variables in JANI
When building JANI models with rewards, we employ an optimization that lifts similar transient edge destination assignments to the edge, so that they only have to be evaluated once for a given edge (and not for each destination). This previously would yield incorrect rewards when the initial (default) value of a transient variable is not zero.

We make the decision when to apply the optimization more conservative to avoid such cases.

Also fixes a minor issue in a sanity check